requirements.txt
tqdm==4.66.1
python-dotenv==1.0.0
langchain==0.3.8
langchain-community==0.0.29
langchain-core==0.3.21
langchain-google-genai==2.0.5
langchain-groq==0.2.1
langchain-openai==0.1.1
langchain-text-splitters==0.3.2